requirements.txt
# ============================================================
# requirements.txt
# Audio ML Environment — Google Colab Mirror
#
# NOTE: This file is a REFERENCE / CI snapshot.
#       For full setup (GPU detection, system libs, correct
#       PyTorch CUDA wheel) always use:
#           python colab_setup.py
#
# For CPU-only environments (CI, testing):
#       pip install -r requirements.txt
# ============================================================

# ── Build tools ─────────────────────────────────────────────
setuptools>=68
wheel
Cython
pybind11
cffi

# ── CRITICAL pin — must stay <2.0 for torch AND tensorflow ──
numpy==1.26.4
scipy==1.13.1
packaging>=23.1
typing-extensions>=4.8

# ── PyTorch (CPU fallback — GPU via colab_setup.py) ─────────
# torch==2.3.1          ← installed by colab_setup.py with correct CUDA wheel
# torchvision==0.18.1
# torchaudio==2.3.1
torchtext==0.18.0
torchmetrics==1.4.0
lightning==2.3.3
timm==1.0.3
einops==0.8.0
opt-einsum==3.3.0

# ── TensorFlow ──────────────────────────────────────────────
tensorflow==2.16.1
keras==3.3.3
tensorflow-hub==0.16.1
tensorflow-datasets==4.9.4
tensorflow-addons==0.23.0

# ── Audio ───────────────────────────────────────────────────
librosa==0.10.2
soundfile==0.12.1
audioread==3.0.1
pydub==0.25.1
noisereduce==3.0.2
resampy==0.4.3
soxr==0.3.7
python-speech-features==0.6
opensmile==2.5.0
pyworld==0.3.4
praat-parselmouth==0.4.3
speechbrain==1.0.0
sounddevice==0.4.7
mutagen==1.47.0
tinytag==1.10.1
ffmpeg-python==0.2.0

# ── NLP ─────────────────────────────────────────────────────
transformers==4.42.4
tokenizers==0.19.1
datasets==2.20.0
accelerate==0.31.0
huggingface-hub==0.23.4
sentence-transformers==3.0.1
nltk==3.8.1
spacy==3.7.5
gensim==4.3.2
sentencepiece==0.2.0
sacremoses==0.1.1

# ── Computer Vision ─────────────────────────────────────────
opencv-python-headless==4.10.0.84
albumentations==1.4.10
Pillow==10.3.0
imageio==2.34.2
scikit-image==0.23.2

# ── Classical ML ────────────────────────────────────────────
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.3
joblib==1.4.2
cloudpickle==3.0.0

# ── Experiment tracking ─────────────────────────────────────
mlflow==2.14.1
wandb==0.17.3
tensorboard==2.16.2
plotly==5.22.0
seaborn==0.13.2
matplotlib==3.9.0
tqdm==4.66.4
rich==13.7.1

# ── Jupyter ─────────────────────────────────────────────────
jupyterlab==4.2.3
notebook==7.2.1
ipywidgets==8.1.3
ipykernel==6.29.4
nbformat==5.10.4
ipython==8.25.0

# ── Utilities ───────────────────────────────────────────────
pandas==2.2.2
h5py==3.11.0
pyarrow==16.1.0
requests==2.32.3
pyyaml==6.0.1
python-dotenv==1.0.1
psutil==6.0.0
numba==0.60.0
llvmlite==0.43.0
